🧀🥨 -> SSDROP-1 & SSDROP-2: Creating form for users to view their relevant NFTs across chains
