+++ client/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@chakra-ui/react": "^2.1.2",
     "@ion-phaser/core": "^1.3.0",
     "@ion-phaser/react": "^1.3.0",
     "@magiclabs/ui": "0.8.0",
@@ -42,6 +43,6 @@
     ]
   },
   "devDependencies": {
-    "framer-motion": "^2.3.0"
+    "framer-motion": "^4.1.17"
   }
 }
