+++ client/src/components/Home.js
@@ -3,9 +3,6 @@ const Home = () => {
     <>
       <h3 className="h3-header">Here's all of our FREE content! 😜</h3>
       <div>♡ FREE.</div>
-      <div>♡ FREE.</div>
-      <div>♡ FREE.</div>
-      <div>♡ FREE.</div>
       <style>{`
         .h3-header {
           font-size: 22px;
