+++ client/src/components/Minigame/GameTabs.js
@@ -1,0 +1,28 @@
+import { Tabs, Tab, TabList, TabPanel, TabPanels } from "@chakra-ui/react";
+
+const GameTabs = () => {
+    return (
+        <>
+            <Tabs>
+                <TabList>
+                    <Tab>One</Tab>
+                    <Tab>Two</Tab>
+                    <Tab>Three</Tab>
+                </TabList>
+                <TabPanels>
+                    <TabPanel>
+                        <p>one!</p>
+                    </TabPanel>
+                    <TabPanel>
+                        <p>two!</p>
+                    </TabPanel>
+                    <TabPanel>
+                        <p>three!</p>
+                    </TabPanel>
+                </TabPanels>
+            </Tabs>
+        </>
+    )
+}
+
+export default GameTabs;
